refactor src/xngin/tq tests, splitting handler tests into its own file
